Refactoring Runtime.Api - VisualizationContext,VisualizationExpression
Moving out of the `Runtime` class.
